% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summary_statistics}
\alias{summary_statistics}
\title{Summary Statistics for Specified Columns}
\usage{
summary_statistics(dataframe, columns, na.rm = TRUE)
}
\arguments{
\item{dataframe}{A dataframe which contains the data for the analysis. The parameter is called "dataframe" so that it is clear for the user that this is the dataset for which the summary statistics will be calculated.}

\item{columns}{A character vector of column names. These are the column/columns for which the summary statistics will be calculated.}

\item{na.rm}{A logical argument. If this is "TRUE", the function removes all "NA" values before calculating the summary statistics.}
}
\value{
A dataframe which is populated by the calculated summary statistics including the columns: mean, median, standard deviation and count of non-NA values of the specified columns. In this dataframe, each row represents a column from the original dataframe and each column represents a particular summary statistic.
}
\description{
This function aids in quick data analysis by calculating the mean, median, standard deviation and count of all non-NA values present in specified columns of a dataframe.
}
\examples{
# Example of using the summary_statistics of penguins from the palmerpenguins dataset.
library(palmerpenguins)
summary_statistics(penguins, columns = c("flipper_length_mm", "body_mass_g"))
}
